Extract of process also used in `Rule` from `Feature`
